examples/CaloINN-calochallenge-generatebranch/requirements.txt
FrEIA==0.2
h5py==3.7.0
matplotlib==3.5.0
numpy==1.24.3
pandas==1.3.5
PyYAML==6.0.1
scikit_learn==1.0.2
scipy==1.13.0
torch==1.13.1
